set up framework for all cpu instructions in match statement

+++ src/and.rs
@@ -40,7 +40,7 @@ pub fn and_indexed_indirect(operand : u8, x_val : u8, pc_reg : &mut u16, accumul
     *cycles_until_next = 6;
 }
 
-pub fn and_indirect_indexed(operand : u16, y_val : u8, pc_reg : &mut u16, accumulator:  &mut u8, status_flags: &mut u8, memory : &RAM, cycles_until_next : &mut u8){
+pub fn and_indirect_indexed(operand : u8, y_val : u8, pc_reg : &mut u16, accumulator:  &mut u8, status_flags: &mut u8, memory : &RAM, cycles_until_next : &mut u8){
     *accumulator = addressing::indirect_indexed(*accumulator, y_val, operand, memory, status_flags, Operation::And);
     *pc_reg += 2;
     *cycles_until_next = 5;
